% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsvm_optimizer.R
\name{lsvm_optimizer}
\alias{lsvm_optimizer}
\title{Lagrangian Support Vector Machine Optimizer}
\usage{
lsvm_optimizer(H, Hinv, q, eps, max.steps, u, beta, ...)
}
\arguments{
\item{H, Hinv, q}{input matrices.}

\item{eps, max.steps}{error and maximum iterations.}

\item{u}{init solution.}

\item{beta}{parameter for speeding up convergence.}

\item{...}{redundant parameters.}
}
\value{
return results list
}
\description{
Lagrangian Support Vector Machine Optimizer
}
\author{
Zhang Jiaqi.
}
